sdks/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reivo"
version = "0.1.0"
description = "Reivo SDK — Cut your AI API costs in half with one line of code."
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "Satoshi Tazoe", email = "hello@reivo.dev" }]
keywords = ["ai", "llm", "openai", "anthropic", "google", "cost", "proxy", "budget"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]

[project.urls]
Homepage = "https://reivo.dev"
Repository = "https://github.com/tazsat0512/reivo"
Documentation = "https://reivo.dev"

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.20.0"]
google = ["google-genai>=1.0.0"]
all = ["openai>=1.0.0", "anthropic>=0.20.0", "google-genai>=1.0.0"]

[tool.hatch.build.targets.wheel]
packages = ["reivo"]
